# backend/JobProcessing.Application/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JobProcessingApi",
    "version": "v1"
  },
  "paths": {
    "/Job/{id}": {
      "get": {
        "tags": [
          "Job"
        ],
        "operationId": "Job_GetJobById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Job": {
      "post": {
        "tags": [
          "Job"
        ],
        "operationId": "Job_CreateJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Job"
        ],
        "operationId": "Job_GetJobs",
        "parameters": [
          {
            "name": "Name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Description",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Price.CurrencyCode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Price.PriceType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/PriceType"
            }
          },
          {
            "name": "Price.MinPrice",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Price.MaxPrice",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "CategoryIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "JobDurationType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/JobDurationType"
            }
          },
          {
            "name": "City",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Job/{id}/document": {
      "post": {
        "tags": [
          "Job"
        ],
        "operationId": "Job_CreateJobDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDocumentCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDocumentCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDocumentCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/Job/GetJobCategories": {
      "get": {
        "tags": [
          "Job"
        ],
        "operationId": "Job_GetJobCategories",
        "parameters": [
          {
            "name": "getCategoriesQuery",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/GetCategoriesQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Job/GetJobLocations": {
      "get": {
        "tags": [
          "Job"
        ],
        "operationId": "Job_GetJobLocations",
        "parameters": [
          {
            "name": "Location",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Address"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Address"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Address"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Address": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "format": "float"
          },
          "longitude": {
            "type": "number",
            "format": "float"
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "addressLine": {
            "type": "string",
            "nullable": true
          },
          "zip": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "categoryName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateJobCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "description": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "highlightedDescription": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "numberOfEmployeesRequired": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "location": {
            "$ref": "#/components/schemas/JobAddressRequest"
          },
          "jobDurationType": {
            "$ref": "#/components/schemas/JobDurationType"
          },
          "price": {
            "$ref": "#/components/schemas/JobPriceRequest"
          },
          "urgency": {
            "$ref": "#/components/schemas/JobUrgency"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "CreateJobDocumentCommand": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "integer",
            "format": "int32"
          },
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "documentFileName": {
            "type": "string",
            "nullable": true
          },
          "isPrimary": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "GetCategoriesQuery": {
        "type": "object",
        "additionalProperties": false
      },
      "JobAddressRequest": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "format": "float"
          },
          "longitude": {
            "type": "number",
            "format": "float"
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "addressLine": {
            "type": "string",
            "nullable": true
          },
          "zip": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "JobDurationType": {
        "enum": [
          "LessThanADay",
          "LessThanAWeek",
          "LessThanAMonth",
          "OneToSixMonths",
          "MoreThanSixMonths"
        ],
        "type": "string"
      },
      "JobPriceRequest": {
        "type": "object",
        "properties": {
          "currencyCode": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "priceType": {
            "$ref": "#/components/schemas/PriceType"
          },
          "minPrice": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "maxPrice": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "JobResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "description": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "highlightedDescription": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "numberOfEmployeesRequired": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "location": {
            "$ref": "#/components/schemas/Address"
          },
          "jobDurationType": {
            "$ref": "#/components/schemas/JobDurationType"
          },
          "price": {
            "$ref": "#/components/schemas/Price"
          },
          "urgency": {
            "$ref": "#/components/schemas/JobUrgency"
          },
          "createDate": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "jobStatus": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "categoryName": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "employerId": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "JobStatus": {
        "enum": [
          "Created",
          "Active",
          "Cancelled",
          "Finished"
        ],
        "type": "string"
      },
      "JobUrgency": {
        "enum": [
          "Normal",
          "Hight",
          "Urgent"
        ],
        "type": "string"
      },
      "Price": {
        "type": "object",
        "properties": {
          "currencyCode": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "priceType": {
            "$ref": "#/components/schemas/PriceType"
          },
          "minPrice": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "maxPrice": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "PriceType": {
        "enum": [
          "Hourly",
          "FixedPrice"
        ],
        "type": "string"
      }
    }
  }
}